Add a prop and a handle click function to the search bar
The prop is used to connect the component to it's parent, which is the
App component, this prop will be a function that we will define in the
parent for passing data from the child to the parent.

We change the handle click function for calling the function that will
be passed down from the parent.